fixed bug in which various exceptions were using a zwlLangException cause overload
- ZwlLangException has contructor overloads with cause that should only be used
  to wrap WebDriverException or RecognitionException because those require
  special handling to build a meningful exceeption and cant be shown to
 useer as is.
Fixed all places that were passing a cause when it wwasnt required as tthose are sent as is and
could have caused problems while interpretting exceeption

+++ src/test/java/com/zylitics/zwl/api/WebdriverTests.java
@@ -397,7 +397,7 @@ public class WebdriverTests {
     if (driver == null) {
       return;
     }
-    int hold = Integer.getInteger("holdWdCloseFor", 0); // in seconds.
+    long hold = Integer.getInteger("holdWdCloseFor", 0); // in seconds.
     if (hold > 0) {
       try {
         Thread.sleep(hold * 1000);
